requirements.txt
numpy
pandas
scipy
scikit-learn
pyarrow
torch
torchvision
torch-geometric
transformers
sentence-transformers
datasets
k-means-constrained
adapt
POT
ortools
matplotlib
tqdm
Pillow
